+++ pkg/internal/ebpf/tcmanager/tcmanager_test.go
@@ -6,13 +6,9 @@
 package tcmanager
 
 import (
-	"bytes"
-	"errors"
-	"os"
 	"testing"
 
 	"github.com/cilium/ebpf"
-	"github.com/cilium/ebpf/link"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/vishvananda/netlink"
@@ -29,102 +25,6 @@ func TestTCXManagerTcxAttachType(t *testing.T) {
 	test(AttachmentIngress, ebpf.AttachTCXIngress)
 }
 
-func isBpfProgLoaded(name string) (bool, error) {
-	id := ebpf.ProgramID(0)
-	var err error
-
-	for {
-		id, err = ebpf.ProgramGetNextID(id)
-		if err != nil {
-			if errors.Is(err, os.ErrNotExist) {
-				break
-			}
-
-			return false, err
-		}
-
-		prog, err := ebpf.NewProgramFromID(id)
-		if err != nil {
-			if errors.Is(err, os.ErrNotExist) {
-				// the program has been closed in the meantime, continue
-				continue
-			}
-
-			return false, err
-		}
-
-		defer prog.Close()
-
-		info, err := prog.Info()
-		if err != nil {
-			return false, err
-		}
-
-		if info.Name == name {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-func isBpfProgLinked(name string) (bool, error) {
-	it := new(link.Iterator)
-
-	for it.Next() {
-		link := it.Take()
-		defer link.Close()
-
-		info, err := link.Info()
-		if err != nil {
-			return false, err
-		}
-
-		prog, err := ebpf.NewProgramFromID(info.Program)
-		if err != nil {
-			return false, err
-		}
-
-		defer prog.Close()
-
-		progInfo, err := prog.Info()
-		if err != nil {
-			return false, err
-		}
-
-		if progInfo.Name == name {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-type Progs struct {
-	Ingress *ebpf.Program `ebpf:"obi_ingress"`
-	Egress  *ebpf.Program `ebpf:"obi_egress"`
-}
-
-func loadProgs(t *testing.T) Progs {
-	reader := bytes.NewReader(tcProgram)
-	assert.NotNil(t, reader)
-
-	spec, err := ebpf.LoadCollectionSpecFromReader(reader)
-	require.NoError(t, err)
-
-	coll, err := ebpf.NewCollection(spec)
-	require.NoError(t, err)
-	assert.NotNil(t, coll)
-
-	var progs Progs
-	err = coll.Assign(&progs)
-	require.NoError(t, err)
-	assert.NotNil(t, progs.Ingress)
-	assert.NotNil(t, progs.Egress)
-
-	return progs
-}
-
 func TestNetlinkManagerNetlinkAttachType(t *testing.T) {
 	test := func(in AttachmentType, out uint32) {
 		r, err := netlinkAttachType(in)
@@ -134,141 +34,4 @@ func TestNetlinkManagerNetlinkAttachType(t *testing.T) {
 
 	test(AttachmentEgress, netlink.HANDLE_MIN_EGRESS)
 	test(AttachmentIngress, netlink.HANDLE_MIN_INGRESS)
-}
-
-func isNetlinkFilterPresent(filterName string, attachType AttachmentType, netManager *netlinkManager) (bool, error) {
-	filterCount := 0
-
-	netManager.mutex.Lock()
-	defer netManager.mutex.Unlock()
-
-	for ifaceIndex := range netManager.interfaces {
-		link, err := netlink.LinkByIndex(ifaceIndex)
-		if err != nil {
-			return false, err
-		}
-
-		aType, err := netlinkAttachType(attachType)
-		if err != nil {
-			return false, err
-		}
-
-		filters, err := netlink.FilterList(link, aType)
-		if err != nil {
-			return false, err
-		}
-
-		for _, filter := range filters {
-			bpfFilter, ok := filter.(*netlink.BpfFilter)
-
-			if !ok {
-				continue
-			}
-
-			if bpfFilter.Name == filterName {
-				filterCount++
-			}
-		}
-	}
-
-	if filterCount == len(netManager.interfaces) {
-		return true, nil
-	}
-
-	return false, nil
-}
-
-/*
-#include <linux/bpf.h>
-#include <bpf/bpf_helpers.h>
-
-char __license[] SEC("license") = "Dual MIT/GPL";
-
-SEC("tc_ingress")
-int obi_ingress(struct __sk_buff *skb) {
-    return 0;
-}
-
-SEC("tc_egress")
-int obi_egress(struct __sk_buff *skb) {
-    return 0;
-}
-
-clang -c -target bpf -O2 -o tc_program.o tc_program.c
-xxd --include tc_program.o > tc_program.c
-*/
-
-var tcProgram = []byte{
-	0x7f, 0x45, 0x4c, 0x46, 0x02, 0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0xf7, 0x00, 0x01, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x58, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00,
-	0x08, 0x00, 0x01, 0x00, 0xb7, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x95, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xb7, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x95, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x44, 0x75, 0x61, 0x6c, 0x20, 0x4d, 0x49, 0x54, 0x2f, 0x47, 0x50, 0x4c,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x4b, 0x00, 0x00, 0x00, 0x04, 0x00, 0xf1, 0xff,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x07, 0x00, 0x00, 0x00, 0x12, 0x00, 0x03, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x1e, 0x00, 0x00, 0x00, 0x12, 0x00, 0x04, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x41, 0x00, 0x00, 0x00, 0x11, 0x00, 0x05, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0d, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x02, 0x03, 0x04, 0x00, 0x2e, 0x74, 0x65, 0x78,
-	0x74, 0x00, 0x6f, 0x62, 0x69, 0x5f, 0x69, 0x6e, 0x67, 0x72, 0x65, 0x73,
-	0x73, 0x00, 0x74, 0x63, 0x5f, 0x69, 0x6e, 0x67, 0x72, 0x65, 0x73, 0x73,
-	0x00, 0x6f, 0x62, 0x69, 0x5f, 0x65, 0x67, 0x72, 0x65, 0x73, 0x73, 0x00,
-	0x74, 0x63, 0x5f, 0x65, 0x67, 0x72, 0x65, 0x73, 0x73, 0x00, 0x2e, 0x6c,
-	0x6c, 0x76, 0x6d, 0x5f, 0x61, 0x64, 0x64, 0x72, 0x73, 0x69, 0x67, 0x00,
-	0x5f, 0x5f, 0x6c, 0x69, 0x63, 0x65, 0x6e, 0x73, 0x65, 0x00, 0x74, 0x63,
-	0x5f, 0x70, 0x72, 0x6f, 0x67, 0x72, 0x61, 0x6d, 0x2e, 0x63, 0x00, 0x2e,
-	0x73, 0x74, 0x72, 0x74, 0x61, 0x62, 0x00, 0x2e, 0x73, 0x79, 0x6d, 0x74,
-	0x61, 0x62, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x58, 0x00, 0x00, 0x00, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0xeb, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x68, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00,
-	0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00,
-	0x01, 0x00, 0x00, 0x00, 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x29, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x06, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x50, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x43, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00,
-	0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x60, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x0d, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x00, 0x00, 0x00,
-	0x03, 0x4c, 0xff, 0x6f, 0x00, 0x00, 0x00, 0x80, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xe8, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x60, 0x00, 0x00, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x70, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x78, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x02, 0x00, 0x00, 0x00,
-	0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00,
 }
